feat(io): refactor io deserialization
fix(datetime): update source format

query(values_source): fallback to parser

Add format settings

Add format settings

query(values_source): improve streaming_load

query(values_source): improve streaming_load

query(values_source): improve streaming_load

+++ query/src/sql/statements/statement_common.rs
@@ -136,7 +136,7 @@ pub fn parse_copy_file_format_options(
     let skip_header = file_format_options
         .get("skip_header")
         .unwrap_or(&"0".to_string())
-        .parse::<i32>()?;
+        .parse::<u64>()?;
 
     // Field delimiter.
     let field_delimiter = file_format_options
